Rework WIS buttons to use imported functions
