+++ js/popup/google_cid.js
@@ -1,4 +1,4 @@
-function googleCID() {
+export function googleCID() {
     chrome.tabs.executeScript({ file: "scripts/jquery-3.2.1.min.js" }, function () {
       chrome.tabs.executeScript({ file: "functions/googleCID.js" });
     });
