+++ js/popup/hub_search.js
@@ -1,3 +1,3 @@
-function hubSearch() {
+export function hubSearch() {
     chrome.tabs.executeScript({ file: "functions/hub-Search.js" });
   }
